backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
celery[redis]
faster-whisper
llama-cpp-python